change the key  q, n, r, h
I went to the "https://gir.st/sol.htm" site in readme file.  In this site,  the key to get out the game is ":quit". And the key  to have new game is ":new". And  the key to start the game again is ":restart". But in game,  the key to get out the game is ":q", not ":quit". ":new", ":restart"  were similar to the situation ahead, too. And I didn't know there was ":help" key. So, I change the source code.  In line 999, there is new code.  The annotations below the new code is the old code. If you do this, you have to enter ":quit" key to get out the game. And you have to  enter ":new" key to have new game.  ":restart", ":help" were similar to the situation ahead, too.