fix find-following
this bug manifested in freecell as the ace of clubs not recognized as
placable on the foundation (from the tableu) when all four free cells
were occupied.
due to us not checking that find_top() found a card at all, we
accidentally checked foundation[-1], which is the fourth free cell.

there are probably more bugs stemming from this misuse of find_top().